refactor(plugin-sanity): move sources to src
