+++ packages/plugins/sanity/package.json
@@ -5,9 +5,9 @@
   "type": "module",
   "exports": {
     ".": {
-      "types": "./index.d.ts",
-      "import": "./index.js",
-      "default": "./index.js"
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js",
+      "default": "./dist/index.js"
     }
   },
   "scripts": {
@@ -22,10 +22,9 @@
     "zod": "^3.25.73"
   },
   "files": [
-    "index.js",
-    "index.d.ts"
+    "dist"
   ],
-  "main": "./index.js",
-  "module": "./index.js",
-  "types": "./index.d.ts"
+  "main": "./dist/index.js",
+  "module": "./dist/index.js",
+  "types": "./dist/index.d.ts"
 }
